Move contest_link_80F57C4 data to src

+++ include/graphics.h
@@ -4877,7 +4877,7 @@ extern const u32 gRouletteNumbersTiles[];
 extern const u32 gRouletteMultiplierTiles[];
 
 // Contest Link
-extern const u32 gUnknown_0858D6B0[];
+extern const u16 gUnknown_0858D6B0[];
 extern const u32 gUnknown_08C19588[];
 extern const u32 gUnknown_08C19EEC[];
 extern const u32 gUnknown_08C1A000[];
